+++ dessia_common/core.py
@@ -28,11 +28,11 @@ from ast import literal_eval
 
 import dessia_common.errors
 from dessia_common.utils.diff import data_eq, diff, dict_hash, list_hash
-from dessia_common.utils.serialization import dict_to_object, serialize_dict_with_pointers, serialize_dict,\
+from dessia_common.utils.serialization import dict_to_object, serialize_dict_with_pointers, serialize_dict, \
     deserialize_argument, serialize
 from dessia_common.utils.types import full_classname, is_sequence, is_bson_valid, TYPES_FROM_STRING
 from dessia_common.utils.copy import deepcopy_value
-from dessia_common.utils.jsonschema import default_dict, jsonschema_from_annotation, JSONSCHEMA_HEADER,\
+from dessia_common.utils.jsonschema import default_dict, jsonschema_from_annotation, JSONSCHEMA_HEADER, \
     set_default_value
 from dessia_common.utils.docstrings import parse_docstring, FAILED_DOCSTRING_PARSING
 from dessia_common.exports import XLSXWriter
@@ -40,6 +40,7 @@ from dessia_common.typings import JsonSerializable
 from dessia_common import templates
 from dessia_common.displays import DisplayObject, DisplaySetting
 from dessia_common.breakdown import attrmethod_getter, get_in_object_from_path
+import dessia_common.files as dcf
 
 _FORBIDDEN_ARGNAMES = ['self', 'cls', 'progress_callback', 'return']
 
@@ -148,7 +149,17 @@ class DessiaObject:
             return self._data_eq(other_object)
         return object.__eq__(self, other_object)
 
-    def _data_eq(self, other_object):
+    def _data_eq_dict(self):
+        """
+        Returns a dict of what to look at for data eq. Keys in non data eq attributes are removed
+        """
+        return {k: v for k, v in self._serializable_dict().items()
+                if k not in self._non_data_eq_attributes + ['package_version', 'name']}
+
+    def _data_eq(self, other_object) -> bool:
+        """
+        Returns if the object is equal to the other object in the sense of data contained in the objects
+        """
         return data_eq(self, other_object)
 
     def _data_hash(self):
@@ -710,7 +721,6 @@ class PhysicalObject(DessiaObject):
         Exports the CAD of the object to step. Works if the class define a custom volmdlr model
         :param filepath: a str representing a filepath
         """
-        self.volmdlr_volume_model().to_step(filepath=filepath)
         return self.volmdlr_volume_model().to_step(filepath=filepath)
 
     def to_step_stream(self, stream):
@@ -718,6 +728,17 @@ class PhysicalObject(DessiaObject):
         Exports the CAD of the object to a stream in the STEP format. Works if the class define a custom volmdlr model
         """
         return self.volmdlr_volume_model().to_step_stream(stream=stream)
+
+    def to_html_stream(self, stream: dcf.StringFile):
+        """
+        Exports the CAD of the object to a stream in the html format.
+        """
+        model = self.volmdlr_volume_model()
+        babylon_data = model.babylon_data()
+        script = model.babylonjs_script(babylon_data)
+        stream.write(script)
+
+        return stream
 
     def to_stl_stream(self, stream):
         """
@@ -753,7 +774,9 @@ class PhysicalObject(DessiaObject):
         formats3d = [{"selector": "step", "extension": "step",
                       "method_name": "to_step_stream", "text": True, "args": {}},
                      {"selector": "stl", "extension": "stl",
-                      "method_name": "to_stl_stream", "text": False, "args": {}}]
+                      "method_name": "to_stl_stream", "text": False, "args": {}},
+                     {"selector": "html", "extension": "html",
+                      "method_name": "to_html_stream", "text": True, "args": {}}]
         formats.extend(formats3d)
         return formats
 
@@ -1353,7 +1376,7 @@ def split_argspecs(argspecs) -> Tuple[int, int]:
 
 
 def get_attribute_names(object_class):
-    attributes = [attribute[0] for attribute in inspect.getmembers(object_class, lambda x:not inspect.isroutine(x))
+    attributes = [attribute[0] for attribute in inspect.getmembers(object_class, lambda x: not inspect.isroutine(x))
                   if not attribute[0].startswith('__')
                   and not attribute[0].endswith('__')
                   and isinstance(attribute[1], (float, int, complex, bool))]